+++ CHANGELOG.md
@@ -2,7 +2,14 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## 1.2.0
+
+- Add the ability to specify environment variables to the container.
+- Upgrade the docker-client to use version 5.0.2
+- Updated travis-ci to test against more versions of Docker.
+
 ## 1.1.0
+> 2016-07-16
 
 - Add a builder to remove the need for inheritance
 - Add a waitForLog directive